Add workspace canaries
An 8-byte random canary value is added between each workspace allocation
that is made by the FD_SCRATCH_ALLOC_APPEND utility macro. The integrity
of the canary value is checked at runtime, such as in the fd_mux
housekeeping code and the after_frag routines of the shred, quic, and
verify tiles.

This change also includes fixes for layout related tests and a bug in
tcache initialization.